requirement.txt
kafka-python==2.0.2
fastapi[all]==0.75.1
python-dotenv==0.20.0
uvicorn~=0.17.0
pydantic~=1.8.2
